+++ lib/bolt/core/src/dep/terraform/install_scripts/mod.rs
@@ -98,7 +98,8 @@ pub async fn gen(
 
 fn gg_traefik_static_config(server: &Server, api_route_token: &str) -> String {
 	let http_provider_endpoint = format!(
-		"http://127.0.0.1:5001/traefik/config/game-guard?token={api_route_token}&region={region}",
+		"http://127.0.0.1:{port}/traefik/config/game-guard?token={api_route_token}&region={region}",
+		port = components::TUNNEL_API_ROUTE_PORT,
 		region = server.region_id
 	);